requirements.txt
Django==4.2.9
psycopg2==2.9.9
django-redis==5.4.0 
celery==5.2.6
redis==4.0.2
flower==2.0.1
djangorestframework==3.14.0
crispy-bootstrap5==2024.2
django-allauth==0.63.3
django-crispy-forms==2.1
django-debug-toolbar==4.4.6
django-mathfilters==1.0.0
environ==1.0
Pillow==10.1.0
pymemcache==4.0.0
python-memcached==1.62
yookassa==3.1.0
python-dotenv==0.20.0
isort==5.13.2